se soluciona problema de ventas para que no intente convertir la key de mercadopago si no es necesario

+++ backend/Mova/src/main/java/com/movauy/mova/repository/finance/CashRegisterRepository.java
@@ -17,7 +17,7 @@ public interface CashRegisterRepository extends JpaRepository<CashRegister, Long
     Optional<CashRegister> findByCloseDateIsNull();
 
     // NUEVO: Buscar caja abierta para un usuario específico
-    Optional<CashRegister> findByCloseDateIsNullAndUser_Id(Integer userId);
+    Optional<CashRegister> findByCloseDateIsNullAndUser_Id(Long userId);
 
     @Query("SELECT c FROM CashRegister c ORDER BY c.openDate DESC")
     List<CashRegister> findAllCashRegisters();
